scripts/requirements.txt
bibtexparser
PyYAML
feedparser
requests
